test: split peripheral tests by fixture
