+++ projects/typescript-tests/build.mjs
@@ -12,18 +12,16 @@ for (const file of ["index.ts", "base.ts", "asserts.ts", "reports.ts"]) {
   cpSync(resolve("node_modules/@siredvin/soteria", file), resolve(soteriaSource, file));
 }
 
-const result = spawnSync(
-  resolve("node_modules/.bin/tstl"),
-  ["-p", "tsconfig.json", "--luaBundle", resolve(output, "peripheralworksgametests.universal_scanner.lua"), "--luaBundleEntry", "src/universal_scanner.ts"],
-  { stdio: "inherit" }
-);
-if (result.status !== 0) process.exit(result.status ?? 1);
-
 for (const peripheral of [
-  "ultimate_sensor", "item_pedestal", "map_pedestal", "display_pedestal",
+  "universal_scanner", "ultimate_sensor", "item_pedestal", "map_pedestal", "display_pedestal",
   "remote_observer", "peripheral_proxy", "reality_forger", "recipe_registry",
   "informative_registry", "statue_workbench", "entity_link", "network_manager",
   "hologram_projector",
 ]) {
-  cpSync(resolve(output, "peripheralworksgametests.universal_scanner.lua"), resolve(output, `peripheralworksgametests.${peripheral}.lua`));
+  const result = spawnSync(
+    resolve("node_modules/.bin/tstl"),
+    ["-p", "tsconfig.json", "--luaBundle", resolve(output, `peripheralworksgametests.${peripheral}.lua`), "--luaBundleEntry", `src/${peripheral}.ts`],
+    { stdio: "inherit" }
+  );
+  if (result.status !== 0) process.exit(result.status ?? 1);
 }
